Rip out the entire setting/mixable system
This system was too ambitious in terms of responsibility. This library
is complex enough as is, managing the entire settings at the same time
was just too much. This commit essentially replaces Mixables with
virtual functions and moves the rendering code to the example. This
should also make it easier to import this library in existing projects